Separate patches into dedicated source files
