+++ CMakeLists.txt
@@ -4,6 +4,19 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -static-libstdc++")
 add_subdirectory(genieutils EXCLUDE_FROM_ALL)
 include_directories(
         "genieutils/include"
+        "."
 )
-add_executable(create-data-mod create-data-mod.cpp patches.cpp ids.h)
+add_executable(
+    create-data-mod
+    create-data-mod.cpp
+    patches/community_games.cpp
+    patches/exploding_villagers.cpp
+    patches/flying_dutchman.cpp
+    patches/kidnap.cpp
+    patches/no_wall.cpp
+    patches/random_costs.cpp
+    patches/duplicate_techs.cpp
+    ids.h
+)
+set_property(TARGET create-data-mod PROPERTY CXX_STANDARD 17)
 target_link_libraries(create-data-mod genieutils)
