+++ patches/community_games.h
@@ -1,0 +1,9 @@
+#ifndef COMMUNITY_GAMES_H
+#define COMMUNITY_GAMES_H
+
+#include "genie/dat/DatFile.h"
+
+
+void configureCommunityGamesMod(genie::DatFile *df);
+
+#endif  // COMMUNITY_GAMES_H
