+++ patches/duplicate_techs.h
@@ -1,0 +1,9 @@
+#ifndef DUPLICATE_TECHS_H
+#define DUPLICATE_TECHS_H
+
+#include "genie/dat/DatFile.h"
+
+
+void duplicateTechs(genie::DatFile *df, int totalCount);
+
+#endif  // DUPLICATE_TECHS_H
