+++ patches/exploding_villagers.h
@@ -1,0 +1,9 @@
+#ifndef EXPLODING_VILLAGERS_H
+#define EXPLODING_VILLAGERS_H
+
+#include "genie/dat/DatFile.h"
+
+
+void configureExplodingVillagers(genie::DatFile *df);
+
+#endif  // EXPLODING_VILLAGERS_H
