+++ patches/flying_dutchman.h
@@ -1,0 +1,9 @@
+#ifndef FLYING_DUTCHMAN_H
+#define FLYING_DUTCHMAN_H
+
+#include "genie/dat/DatFile.h"
+
+
+void makeTransportShipsFly(genie::DatFile *df);
+
+#endif  // FLYING_DUTCHMAN_H
