+++ patches/kidnap.h
@@ -1,0 +1,9 @@
+#ifndef KIDNAP_H
+#define KIDNAP_H
+
+#include "genie/dat/DatFile.h"
+
+
+void configureKidnap(genie::DatFile *df);
+
+#endif  // KIDNAP_H
