+++ patches/no_wall.h
@@ -1,0 +1,9 @@
+#ifndef NO_WALL_H
+#define NO_WALL_H
+
+#include "genie/dat/DatFile.h"
+
+
+void disableWalls(genie::DatFile *df);
+
+#endif  // NO_WALL_H
