+++ patches/random_costs.h
@@ -1,0 +1,11 @@
+#ifndef RANDOM_COSTS_H
+#define RANDOM_COSTS_H
+
+#include "genie/dat/DatFile.h"
+
+
+void jumbleCosts(genie::DatFile *df);
+void jumbleUnitCosts(genie::DatFile *df);
+void jumbleTechCosts(genie::DatFile *df);
+
+#endif  // RANDOM_COSTS_H
